Add support for php-fpm on Apache.
This adds support for a new directive for the `server.sapi`
allowing you to specify fpm and mod_php (defaulting to mod_php
for now for backward compatibility, future auto-detection would
be nice). If you set `server.sapi` to `fpm` you can then also
use the new `server.fpm_url` to connect to the host and port
appropriate to your fpm instance.