Prevent fetcher-create from fetching info.
When running fetcher-create you are creating a new site that the
fetcher info system does not know about. Despite that the shared
code path betweed fetcher-fetch and fetcher-create would trigger
an info fetch which would fail (because the site is new) or get
data from an existing site by the same name despite the fact that
you were explicitly trying to create a new one.

This is not a particularly elegant solution but it does prevent
that by adding an explicit flag to the fetcher_get_site function
that instantiates a Site class object.